Remove circular dependency between nested and core classes.
